Refactoring: move bucketChunkReader to dedicated file
