+++ pkg/storegateway/bucket.go
@@ -6,10 +6,8 @@
 package storegateway
 
 import (
-	"bufio"
 	"bytes"
 	"context"
-	"encoding/binary"
 	"fmt"
 	"io"
 	"math"
@@ -44,7 +42,6 @@ import (
 
 	"github.com/grafana/mimir/pkg/mimirpb"
 	"github.com/grafana/mimir/pkg/storage/sharding"
-	mimir_tsdb "github.com/grafana/mimir/pkg/storage/tsdb"
 	"github.com/grafana/mimir/pkg/storage/tsdb/block"
 	"github.com/grafana/mimir/pkg/storage/tsdb/metadata"
 	"github.com/grafana/mimir/pkg/storegateway/hintspb"
@@ -1753,242 +1750,6 @@ func decodeSeriesForTime(b []byte, lset *[]symbolizedLabel, chks *[]chunks.Meta,
 		mint = maxt
 	}
 	return len(*chks) > 0, d.Err()
-}
-
-type loadIdx struct {
-	offset uint32
-	// Indices, not actual entries and chunks.
-	seriesEntry int
-	chunk       int
-}
-
-type bucketChunkReader struct {
-	ctx   context.Context
-	block *bucketBlock
-
-	toLoad [][]loadIdx
-
-	// Mutex protects access to following fields, when updated from chunks-loading goroutines.
-	// After chunks are loaded, mutex is no longer used.
-	mtx        sync.Mutex
-	stats      *queryStats
-	chunkBytes []*[]byte // Byte slice to return to the chunk pool on close.
-}
-
-func newBucketChunkReader(ctx context.Context, block *bucketBlock) *bucketChunkReader {
-	return &bucketChunkReader{
-		ctx:    ctx,
-		block:  block,
-		stats:  &queryStats{},
-		toLoad: make([][]loadIdx, len(block.chunkObjs)),
-	}
-}
-
-func (r *bucketChunkReader) Close() error {
-	r.block.pendingReaders.Done()
-
-	for _, b := range r.chunkBytes {
-		r.block.chunkPool.Put(b)
-	}
-	return nil
-}
-
-// addLoad adds the chunk with id to the data set to be fetched.
-// Chunk will be fetched and saved to res[seriesEntry][chunk] upon r.load(res, <...>) call.
-func (r *bucketChunkReader) addLoad(id chunks.ChunkRef, seriesEntry, chunk int) error {
-	var (
-		seq = int(id >> 32)
-		off = uint32(id)
-	)
-	if seq >= len(r.toLoad) {
-		return errors.Errorf("reference sequence %d out of range", seq)
-	}
-	r.toLoad[seq] = append(r.toLoad[seq], loadIdx{off, seriesEntry, chunk})
-	return nil
-}
-
-// load loads all added chunks and saves resulting aggrs to res.
-func (r *bucketChunkReader) load(res []seriesEntry, aggrs []storepb.Aggr) error {
-	g, ctx := errgroup.WithContext(r.ctx)
-
-	for seq, pIdxs := range r.toLoad {
-		sort.Slice(pIdxs, func(i, j int) bool {
-			return pIdxs[i].offset < pIdxs[j].offset
-		})
-		parts := r.block.partitioner.Partition(len(pIdxs), func(i int) (start, end uint64) {
-			return uint64(pIdxs[i].offset), uint64(pIdxs[i].offset) + mimir_tsdb.EstimatedMaxChunkSize
-		})
-
-		for _, p := range parts {
-			seq := seq
-			p := p
-			indices := pIdxs[p.ElemRng[0]:p.ElemRng[1]]
-			g.Go(func() error {
-				return r.loadChunks(ctx, res, aggrs, seq, p, indices)
-			})
-		}
-	}
-	return g.Wait()
-}
-
-// loadChunks will read range [start, end] from the segment file with sequence number seq.
-// This data range covers chunks starting at supplied offsets.
-func (r *bucketChunkReader) loadChunks(ctx context.Context, res []seriesEntry, aggrs []storepb.Aggr, seq int, part Part, pIdxs []loadIdx) error {
-	fetchBegin := time.Now()
-
-	// Get a reader for the required range.
-	reader, err := r.block.chunkRangeReader(ctx, seq, int64(part.Start), int64(part.End-part.Start))
-	if err != nil {
-		return errors.Wrap(err, "get range reader")
-	}
-	defer runutil.CloseWithLogOnErr(r.block.logger, reader, "readChunkRange close range reader")
-	bufReader := bufio.NewReaderSize(reader, mimir_tsdb.EstimatedMaxChunkSize)
-
-	locked := true
-	r.mtx.Lock()
-
-	defer func() {
-		if locked {
-			r.mtx.Unlock()
-		}
-	}()
-
-	r.stats.chunksFetchCount++
-	r.stats.chunksFetched += len(pIdxs)
-	r.stats.chunksFetchDurationSum += time.Since(fetchBegin)
-	r.stats.chunksFetchedSizeSum += int(part.End - part.Start)
-
-	var (
-		buf        = make([]byte, mimir_tsdb.EstimatedMaxChunkSize)
-		readOffset = int(pIdxs[0].offset)
-
-		// Save a few allocations.
-		written  int64
-		diff     uint32
-		chunkLen int
-		n        int
-	)
-
-	for i, pIdx := range pIdxs {
-		// Fast forward range reader to the next chunk start in case of sparse (for our purposes) byte range.
-		for readOffset < int(pIdx.offset) {
-			written, err = io.CopyN(io.Discard, bufReader, int64(pIdx.offset)-int64(readOffset))
-			if err != nil {
-				return errors.Wrap(err, "fast forward range reader")
-			}
-			readOffset += int(written)
-		}
-		// Presume chunk length to be reasonably large for common use cases.
-		// However, declaration for EstimatedMaxChunkSize warns us some chunks could be larger in some rare cases.
-		// This is handled further down below.
-		chunkLen = mimir_tsdb.EstimatedMaxChunkSize
-		if i+1 < len(pIdxs) {
-			if diff = pIdxs[i+1].offset - pIdx.offset; int(diff) < chunkLen {
-				chunkLen = int(diff)
-			}
-		}
-		cb := buf[:chunkLen]
-		n, err = io.ReadFull(bufReader, cb)
-		readOffset += n
-		// Unexpected EOF for last chunk could be a valid case. Any other errors are definitely real.
-		if err != nil && !(errors.Is(err, io.ErrUnexpectedEOF) && i == len(pIdxs)-1) {
-			return errors.Wrapf(err, "read range for seq %d offset %x", seq, pIdx.offset)
-		}
-
-		chunkDataLen, n := binary.Uvarint(cb)
-		if n < 1 {
-			return errors.New("reading chunk length failed")
-		}
-
-		// Chunk length is n (number of bytes used to encode chunk data), 1 for chunk encoding and chunkDataLen for actual chunk data.
-		// There is also crc32 after the chunk, but we ignore that.
-		chunkLen = n + 1 + int(chunkDataLen)
-		if chunkLen <= len(cb) {
-			err = populateChunk(&(res[pIdx.seriesEntry].chks[pIdx.chunk]), rawChunk(cb[n:chunkLen]), aggrs, r.save)
-			if err != nil {
-				return errors.Wrap(err, "populate chunk")
-			}
-			r.stats.chunksTouched++
-			r.stats.chunksTouchedSizeSum += int(chunkDataLen)
-			continue
-		}
-
-		// If we didn't fetch enough data for the chunk, fetch more.
-		r.mtx.Unlock()
-		locked = false
-
-		fetchBegin = time.Now()
-
-		// Read entire chunk into new buffer.
-		// TODO: readChunkRange call could be avoided for any chunk but last in this particular part.
-		nb, err := r.block.readChunkRange(ctx, seq, int64(pIdx.offset), int64(chunkLen), []byteRange{{offset: 0, length: chunkLen}})
-		if err != nil {
-			return errors.Wrapf(err, "preloaded chunk too small, expecting %d, and failed to fetch full chunk", chunkLen)
-		}
-		if len(*nb) != chunkLen {
-			return errors.Errorf("preloaded chunk too small, expecting %d", chunkLen)
-		}
-
-		r.mtx.Lock()
-		locked = true
-
-		r.stats.chunksFetchCount++
-		r.stats.chunksFetchDurationSum += time.Since(fetchBegin)
-		r.stats.chunksFetchedSizeSum += len(*nb)
-		err = populateChunk(&(res[pIdx.seriesEntry].chks[pIdx.chunk]), rawChunk((*nb)[n:]), aggrs, r.save)
-		if err != nil {
-			r.block.chunkPool.Put(nb)
-			return errors.Wrap(err, "populate chunk")
-		}
-		r.stats.chunksTouched++
-		r.stats.chunksTouchedSizeSum += int(chunkDataLen)
-
-		r.block.chunkPool.Put(nb)
-	}
-	return nil
-}
-
-// save saves a copy of b's payload to a memory pool of its own and returns a new byte slice referencing said copy.
-// Returned slice becomes invalid once r.block.chunkPool.Put() is called.
-func (r *bucketChunkReader) save(b []byte) ([]byte, error) {
-	// Ensure we never grow slab beyond original capacity.
-	if len(r.chunkBytes) == 0 ||
-		cap(*r.chunkBytes[len(r.chunkBytes)-1])-len(*r.chunkBytes[len(r.chunkBytes)-1]) < len(b) {
-		s, err := r.block.chunkPool.Get(len(b))
-		if err != nil {
-			return nil, errors.Wrap(err, "allocate chunk bytes")
-		}
-		r.chunkBytes = append(r.chunkBytes, s)
-	}
-	slab := r.chunkBytes[len(r.chunkBytes)-1]
-	*slab = append(*slab, b...)
-	return (*slab)[len(*slab)-len(b):], nil
-}
-
-// rawChunk is a helper type that wraps a chunk's raw bytes and implements the chunkenc.Chunk
-// interface over it.
-// It is used to Store API responses which don't need to introspect and validate the chunk's contents.
-type rawChunk []byte
-
-func (b rawChunk) Encoding() chunkenc.Encoding {
-	return chunkenc.Encoding(b[0])
-}
-
-func (b rawChunk) Bytes() []byte {
-	return b[1:]
-}
-func (b rawChunk) Compact() {}
-
-func (b rawChunk) Iterator(_ chunkenc.Iterator) chunkenc.Iterator {
-	panic("invalid call")
-}
-
-func (b rawChunk) Appender() (chunkenc.Appender, error) {
-	panic("invalid call")
-}
-
-func (b rawChunk) NumSamples() int {
-	panic("invalid call")
 }
 
 // NewDefaultChunkBytesPool returns a chunk bytes pool with default settings.
